requirements.txt
Flask>=3.0.0
psycopg2-binary>=2.9.0
pywebview>=5.0
openpyxl>=3.0.0
